Fix for Collections naming/allocation.
Fixed the issue of collections being named for the temp files instead of the original input file. Also set up the collection better; can now set a specific name, and it will create a new collection if the collection it wants exists but is excluded from the view layer. (Previously it just errored, because it passed the 'test' of knowing the collection existed. I could make it automatically un-exclude the collection but that feels a bit too overbearing,  so I'm saving that for the UI panel.