Minor updates to both scripts, no major function change.
Minor update to init's custom bones, and some work on the obj component identification in the the importer script. The importer script is now messy as hell, but I need to take a break and cull it back later. Committing it anyway so I can't accidentally strip something important later. Some good among the chaff, I think.